generate link of urls for now cuz google tricky bad man
