+++ pages/index.tsx
@@ -1,9 +1,12 @@
 import Head from 'next/head'
 import styles from '../styles/Home.module.css'
+import FralexLife from './_app'
 
 export default function Home() {
   return (
     <div className={styles.container}>
+      <FralexLife></FralexLife>
+
       <Head>
         <title>Create Next App</title>
         <link rel="icon" href="/favicon.ico" />
